+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/common/markdown/ChatMarkdown.kt
@@ -2,6 +2,11 @@ package io.getstream.chat.android.ui.common.markdown
 
 import android.widget.TextView
 
+import kotlin.Deprecated
+
+@Deprecated(message = "ChatMarkdown is deprecated in favour of ChatMessageTextTransformer. " +
+    "Use ChatMessageTextTransformer for text ui transformations."
+)
 public fun interface ChatMarkdown {
     public fun setText(textView: TextView, text: String)
 }